Created the coupon function

+++ controllers/CouponController.php
@@ -2,12 +2,15 @@
 
 namespace app\controllers;
 
+use app\models\ParticipantHasCoupon;
+use app\models\User;
 use Yii;
 use app\models\Coupon;
-use yii\data\ActiveDataProvider;
+use app\models\CouponSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\web\ForbiddenHttpException;
 
 /**
  * CouponController implements the CRUD actions for Coupon model.
@@ -30,15 +33,25 @@ class CouponController extends Controller
      * Lists all Coupon models.
      * @return mixed
      */
-    public function actionIndex()
+    public function actionIndex($apply = 0)
     {
-        $dataProvider = new ActiveDataProvider([
-            'query' => Coupon::find(),
-        ]);
+        if(Yii::$app->user->can('couponIndex')) {
+            $searchModel = new CouponSearch();
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        return $this->render('index', [
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+                'apply' => $apply,
+                'coupons' => Coupon::find()->asArray()->all(),
+            ]);
+        } else {
+            if(Yii::$app->user->isGuest) {
+                Yii::$app->user->loginRequired();
+            } else {
+                throw new ForbiddenHttpException(Yii::t('yii', 'You are not allowed to perform this action.'));
+            }
+        }
     }
 
     /**
@@ -48,9 +61,42 @@ class CouponController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if(Yii::$app->user->can('couponView')) {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        } else {
+            if(Yii::$app->user->isGuest) {
+                Yii::$app->user->loginRequired();
+            } else {
+                throw new ForbiddenHttpException(Yii::t('yii', 'You are not allowed to perform this action.'));
+            }
+        }
+    }
+
+    public function actionApply($id) {
+        if(Yii::$app->user->can('couponGet')) {
+            $model = $this->findModel($id);
+            $score = User::getScore();
+            if($score >= $model->requireScore) {
+                User::reduceScore($model->requireScore);
+                $hasCoupon = new ParticipantHasCoupon();
+                $hasCoupon->Participant_id = Yii::$app->user->id;
+                $hasCoupon->coupon_id = $id;
+                $hasCoupon->is_read = 0;
+                $hasCoupon->is_used = 0;
+                if($hasCoupon->save()) {
+                    $this->redirect(['index', 'apply' => 1]);
+                }
+            }
+            $this->redirect(['index', 'apply' => 2]);
+        } else {
+            if(Yii::$app->user->isGuest) {
+                Yii::$app->user->loginRequired();
+            } else {
+                throw new ForbiddenHttpException(Yii::t('yii', 'You are not allowed to perform this action.'));
+            }
+        }
     }
 
     /**
@@ -60,14 +106,22 @@ class CouponController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Coupon();
-
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+        if(Yii::$app->user->can('couponCreate')) {
+            $model = new Coupon();
+            $model->lastModifyDatetime = date("Y-m-d H:i:s");
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                return $this->render('create', [
+                    'model' => $model,
+                ]);
+            }
         } else {
-            return $this->render('create', [
-                'model' => $model,
-            ]);
+            if(Yii::$app->user->isGuest) {
+                Yii::$app->user->loginRequired();
+            } else {
+                throw new ForbiddenHttpException(Yii::t('yii', 'You are not allowed to perform this action.'));
+            }
         }
     }
 
@@ -79,14 +133,23 @@ class CouponController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if(Yii::$app->user->can('couponUpdate')) {
+            $model = $this->findModel($id);
+            $model->lastModifyDatetime = date("Y-m-d H:i:s");
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                return $this->render('update', [
+                    'model' => $model,
+                ]);
+            }
         } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
+            if(Yii::$app->user->isGuest) {
+                Yii::$app->user->loginRequired();
+            } else {
+                throw new ForbiddenHttpException(Yii::t('yii', 'You are not allowed to perform this action.'));
+            }
         }
     }
 
@@ -98,9 +161,17 @@ class CouponController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(Yii::$app->user->can('couponDelete')) {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        } else {
+            if(Yii::$app->user->isGuest) {
+                Yii::$app->user->loginRequired();
+            } else {
+                throw new ForbiddenHttpException(Yii::t('yii', 'You are not allowed to perform this action.'));
+            }
+        }
     }
 
     /**
